feature: alimentando NewsLetterScreen

+++ app/src/main/java/br/com/fiap/satoshi/model/CryptoSustainable.kt
@@ -10,5 +10,5 @@ data class CryptoSustainable (
     var image: String = "",
     @SerializedName("esg_score") var esgScore: Int = 0,
     @SerializedName("current_price") var currentPrice: Double = 0.0,
-    @SerializedName("market_cap") var marketCap: Int = 0
+    @SerializedName("market_cap") var marketCap: Long = 0
 )
